requirements.txt
pandas>=2.0.0
pyodbc>=4.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
